Uploading Code and CSV File
Code to create the final model with ~87% accuracy [ibrahim (model training).py], evaluate/validate the model [evaluator.py], and house a GUI for practical implementation and demonstration [cancer_detector.py]. This is accompanied by annotations.csv for dataframing. This code cannot be run without the dataset, a 333 MB file consisting of 3,152 images 224x224 image from the MHIST dataset [https://bmirds.github.io/MHIST/] which is too large to upload directly to this repository.